Adding time object
Time object is used for any time-based dependent manipulations such as
animations. An object can subscribe for a time event and get an update
once a time is elapsed with a callback of (t, dt) where t is the current
time and dt is the time passed. For accuracy of simulations the time
object keeps dt constant and is defined once the time object is created.