Added DreamSetup to install Dream
DreamSetup project is added to the solution. Once Dream is built in
Release, a manual call to DreamOS\Scripts\make_production_build.py should
be made to create a Production folder out of the Release build (there was
no way to put it in a pre-build script).
Afterwards building DreamSetup will make the installer, DreamSetup.msi.

Also Dream icons were added to the desktop shortcut, app exe and window.