Added CommandPrompt
CmdPrompt allows the control of object using a command-line string.
To support control of an object using CmdPrompt, follow DreamGarage or DreamConsole inheritence of Subscriber<CmdPromptEvent>.
An Object is allowed to register to a method in CmdPrompt which is a string.
Other arguments will be passed to the object who will get notified when CmdPrompt is triggered.
Now a command line is passed through DreamConsole, and in the future could be sent
by an outside process via ipc.