Make CloudController to run in a separate process
Cloud controller runs in a separate process.
for WebRTC reasons, a message loop should run in the thread. With WebRTC
running in the main thread it was stealing window messaging and thus
making the window not responssive to events. In the seperate thread
approach, the cloud runs clean and doesn't interfere.
Some more works needs to be done to the design of the message loop in
other OS.